Tests/LibWeb/Text/expected/indexeddb-commit-event-ordering.txt

setup.upgradeneeded
setup.success
tx.commit() called
putA.success
putB.success
tx.complete
DONE
